+++ NEWS.md
@@ -1,6 +1,10 @@
-# teal.logger 0.1.3.9011
+# teal.logger 0.2.0.9001
 
+# teal.logger 0.2.0
+
+* New function `register_handlers` to register global handlers for logging messages, warnings and errors.
 * Specified minimal version of package dependencies.
+* Update installation instructions in `README`.
 
 # teal.logger 0.1.3